split phrase matcher samples to two files
