+++ samples/snippets/create_phrase_matcher_all_of.py
@@ -13,53 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# [START contactcenterinsights_create_phrase_matcher_phone_or_cellphone]
+# [START contactcenterinsights_create_phrase_matcher_all_of]
 from google.cloud import contact_center_insights_v1
 
 
-def create_phrase_matcher_phone_or_cellphone(
-    project_id: str,
-) -> contact_center_insights_v1.PhraseMatcher:
-    # Construct a parent resource.
-    parent = contact_center_insights_v1.ContactCenterInsightsClient.common_location_path(
-        project_id, "us-central1"
-    )
-
-    # Construct a phrase matcher that matches any of its rule groups.
-    phrase_matcher = contact_center_insights_v1.PhraseMatcher()
-    phrase_matcher.display_name = "PHONE_SERVICE"
-    phrase_matcher.type_ = (
-        contact_center_insights_v1.PhraseMatcher.PhraseMatcherType.ANY_OF
-    )
-    phrase_matcher.active = True
-
-    # Construct a rule group to match any of the words "PHONE" or "CELLPHONE", ignoring case sensitivity.
-    rule_group = contact_center_insights_v1.PhraseMatchRuleGroup()
-    rule_group.type_ = (
-        contact_center_insights_v1.PhraseMatchRuleGroup.PhraseMatchRuleGroupType.ANY_OF
-    )
-
-    for word in ["PHONE", "CELLPHONE"]:
-        rule = contact_center_insights_v1.PhraseMatchRule()
-        rule.query = word
-        rule.config.exact_match_config = contact_center_insights_v1.ExactMatchConfig()
-        rule_group.phrase_match_rules.append(rule)
-    phrase_matcher.phrase_match_rule_groups.append(rule_group)
-
-    # Call the Insights client to create a phrase matcher.
-    insights_client = contact_center_insights_v1.ContactCenterInsightsClient()
-    phrase_matcher = insights_client.create_phrase_matcher(
-        parent=parent, phrase_matcher=phrase_matcher
-    )
-
-    print(f"Created a phrase matcher named {phrase_matcher.name}")
-    return phrase_matcher
-
-
-# [END contactcenterinsights_create_phrase_matcher_phone_or_cellphone]
-
-# [START contactcenterinsights_create_phrase_matcher_phone_or_cellphone_not_shipping_or_delivery]
-def create_phrase_matcher_phone_or_cellphone_not_shipping_or_delivery(
+def create_phrase_matcher_all_of(
     project_id: str,
 ) -> contact_center_insights_v1.PhraseMatcher:
     # Construct a parent resource.
@@ -75,7 +33,7 @@ def create_phrase_matcher_phone_or_cellphone_not_shipping_or_delivery(
     )
     phrase_matcher.active = True
 
-    # Construct a rule group to match any of the words "PHONE" or "CELLPHONE", ignoring case sensitivity.
+    # Construct a rule group to match the word "PHONE" or "CELLPHONE", ignoring case sensitivity.
     rule_group1 = contact_center_insights_v1.PhraseMatchRuleGroup()
     rule_group1.type_ = (
         contact_center_insights_v1.PhraseMatchRuleGroup.PhraseMatchRuleGroupType.ANY_OF
@@ -88,7 +46,7 @@ def create_phrase_matcher_phone_or_cellphone_not_shipping_or_delivery(
         rule_group1.phrase_match_rules.append(rule)
     phrase_matcher.phrase_match_rule_groups.append(rule_group1)
 
-    # Construct another rule group to not match any of the words "SHIPPING" or "DELIVERY", ignoring case sensitivity.
+    # Construct another rule group to not match the word "SHIPPING" or "DELIVERY", ignoring case sensitivity.
     rule_group2 = contact_center_insights_v1.PhraseMatchRuleGroup()
     rule_group2.type_ = (
         contact_center_insights_v1.PhraseMatchRuleGroup.PhraseMatchRuleGroupType.ALL_OF
@@ -108,8 +66,8 @@ def create_phrase_matcher_phone_or_cellphone_not_shipping_or_delivery(
         parent=parent, phrase_matcher=phrase_matcher
     )
 
-    print(f"Created a phrase matcher named {phrase_matcher.name}")
+    print(f"Created {phrase_matcher.name}")
     return phrase_matcher
 
 
-# [END contactcenterinsights_create_phrase_matcher_phone_or_cellphone_not_shipping_or_delivery]
+# [END contactcenterinsights_create_phrase_matcher_all_of]
